requirements.txt
tabula-py>=2.0.0
